added linking using makefile. now installs
