+++ R/RcppExports.R
@@ -1,36 +1,8 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-dloglik_likelihood_stratify <- function(knot, facility, delta, z, b_spline, theta) {
-    .Call(`_TimeVarying_dloglik_likelihood_stratify`, knot, facility, delta, z, b_spline, theta)
-}
-
-ddloglik <- function(knot, facility, delta, z, b_spline, theta, number_facility) {
-    .Call(`_TimeVarying_ddloglik`, knot, facility, delta, z, b_spline, theta, number_facility)
-}
-
-dloglik_likelihood_gradient <- function(knot, facility, delta, z, b_spline, theta) {
-    .Call(`_TimeVarying_dloglik_likelihood_gradient`, knot, facility, delta, z, b_spline, theta)
-}
-
-NRloop <- function(knot, facility, delta, z, b_spline, theta, M_stop, number_facility, rate, tol) {
-    .Call(`_TimeVarying_NRloop`, knot, facility, delta, z, b_spline, theta, M_stop, number_facility, rate, tol)
-}
-
 sign_num <- function(x) {
     .Call(`_TimeVarying_sign_num`, x)
-}
-
-ddloglik_stratify <- function(knot, facility, z, delta, b_spline, theta, number_facility = 1L) {
-    .Call(`_TimeVarying_ddloglik_stratify`, knot, facility, z, delta, b_spline, theta, number_facility)
-}
-
-GDboost_stratify <- function(knot, rate, facility, delta, z, b_spline, theta) {
-    .Call(`_TimeVarying_GDboost_stratify`, knot, rate, facility, delta, z, b_spline, theta)
-}
-
-GDboost_stratify_no_GD2 <- function(knot, rate, facility, delta, z, b_spline, theta) {
-    .Call(`_TimeVarying_GDboost_stratify_no_GD2`, knot, rate, facility, delta, z, b_spline, theta)
 }
 
 dloglik_likelihood_stratify <- function(knot, facility, delta, z, b_spline, theta) {
@@ -49,15 +21,11 @@ dloglik_likelihood_gradient <- function(knot, facility, delta, z, b_spline, thet
     .Call(`_TimeVarying_dloglik_likelihood_gradient`, knot, facility, delta, z, b_spline, theta)
 }
 
-dloglik_likelihood <- function(par, knot, facility, delta, z, b_spline) {
-    .Call(`_TimeVarying_dloglik_likelihood`, par, knot, facility, delta, z, b_spline)
-}
-
-ddloglik <- function(knot, facility, delta, z, b_spline, theta, number_facility = 1L) {
+ddloglik <- function(knot, facility, delta, z, b_spline, theta, number_facility) {
     .Call(`_TimeVarying_ddloglik`, knot, facility, delta, z, b_spline, theta, number_facility)
 }
 
-stratify <- function(knot, tol, facility, delta, z, b_spline, max_iteration = 5000L) {
-    .Call(`_TimeVarying_stratify`, knot, tol, facility, delta, z, b_spline, max_iteration)
+NRloop <- function(knot, facility, delta, z, b_spline, theta, M_stop, number_facility, rate, tol) {
+    .Call(`_TimeVarying_NRloop`, knot, facility, delta, z, b_spline, theta, M_stop, number_facility, rate, tol)
 }
