+++ src/Makevars.win
@@ -1,0 +1,1 @@
+PKG_LIBS = $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)
